refactor(demo): 将项目改为微服务格式
