+++ app/demo/internal/service/demo.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	v1 "github.com/CyanPigeon/kratos-template/api/demo/v1"
-	"github.com/CyanPigeon/kratos-template/internal/biz"
+	"github.com/CyanPigeon/kratos-template/app/demo/internal/biz"
 	"github.com/go-kratos/kratos/v2/log"
 )
